........
  r4155 | dsdale | 2007-11-07 19:09:17 -0500 (Wed, 07 Nov 2007) | 2 lines
  
  move configobj.py to lib/, install in site-packages only if required
........
  r4156 | dsdale | 2007-11-07 19:24:57 -0500 (Wed, 07 Nov 2007) | 3 lines
  
  update enthought package to version 2.6b1, stripped of setuptools
  installed in site-packages, if not already present
........
  r4160 | mdboom | 2007-11-08 09:31:15 -0500 (Thu, 08 Nov 2007) | 2 lines
  
  Throw in dummy characters for symbols not in the Bakoma fonts.
........
  r4167 | dsdale | 2007-11-08 18:25:44 -0500 (Thu, 08 Nov 2007) | 3 lines
  
  move pyparsing from lib/matplotlib/ to lib/
  install pyparsing only if not already available
........
  r4168 | jdh2358 | 2007-11-08 18:29:46 -0500 (Thu, 08 Nov 2007) | 2 lines
  
  added recarray utils module
........
  r4169 | dsdale | 2007-11-08 19:19:45 -0500 (Thu, 08 Nov 2007) | 2 lines
  
  updated pyparsing to version 1.4.8
........
  r4170 | dsdale | 2007-11-08 19:22:03 -0500 (Thu, 08 Nov 2007) | 3 lines
  
  added checks that also print version numbers for 
  pyparsing, pytz, dateutil, configobj
........
  r4171 | dsdale | 2007-11-08 19:31:48 -0500 (Thu, 08 Nov 2007) | 2 lines
  
  note pyparsing change to API_CHANGES
........
  r4172 | dsdale | 2007-11-08 20:37:02 -0500 (Thu, 08 Nov 2007) | 2 lines
  
  remove old version of pytz
........
  r4173 | dsdale | 2007-11-08 20:40:54 -0500 (Thu, 08 Nov 2007) | 2 lines
  
  added pytz-2007g
........
  r4174 | dsdale | 2007-11-08 20:43:18 -0500 (Thu, 08 Nov 2007) | 2 lines
  
  fixed bug in dateutil version check
........
  r4175 | jdh2358 | 2007-11-08 21:55:47 -0500 (Thu, 08 Nov 2007) | 2 lines
  
  reverted rec utils to mlab
........
  r4176 | efiring | 2007-11-09 02:25:08 -0500 (Fri, 09 Nov 2007) | 2 lines
  
  Pylab uses numpy instead of oldnumeric
........
  r4177 | dsdale | 2007-11-09 08:08:48 -0500 (Fri, 09 Nov 2007) | 2 lines
  
  move pyparsing back into mpl namespace
........
  r4178 | mdboom | 2007-11-09 08:10:12 -0500 (Fri, 09 Nov 2007) | 3 lines
  
  Avoid annoying Qt4 messages when mpl raises an exception.  (Thanks to
  Martin Teichmann in patch 1828813)
........
  r4179 | mdboom | 2007-11-09 08:19:38 -0500 (Fri, 09 Nov 2007) | 2 lines
  
  Fix font caching bug on OSX
........
  r4180 | dsdale | 2007-11-09 08:20:10 -0500 (Fri, 09 Nov 2007) | 3 lines
  
  committed Martin Teichmann's patch 1828813 to fix qt4 error messages
  related to QPainter
........
  r4181 | mdboom | 2007-11-09 08:47:40 -0500 (Fri, 09 Nov 2007) | 3 lines
  
  Remove duplicate line caused by myself and Darren committing the same
  patch and virtually the same time.
........
  r4182 | dsdale | 2007-11-09 08:49:54 -0500 (Fri, 09 Nov 2007) | 2 lines
  
  updated CHANGELOG and API_CHANGES
........
  r4183 | dsdale | 2007-11-09 09:24:41 -0500 (Fri, 09 Nov 2007) | 2 lines
  
  updated dependency report during build process
........
  r4187 | jdh2358 | 2007-11-09 11:42:55 -0500 (Fri, 09 Nov 2007) | 2 lines
  
  added face and edge color = 'None' support to patches
........
  r4188 | jdh2358 | 2007-11-09 11:43:38 -0500 (Fri, 09 Nov 2007) | 2 lines
  
  removed unneeded draw command from tkagg
........

svn path=/branches/transforms/; revision=4192

+++ examples/image_demo2.py
@@ -3,7 +3,7 @@ from pylab import *
 
 w, h = 512, 512
 s = file('data/ct.raw', 'rb').read()
-A = fromstring(s, UInt16).astype(Float)
+A = fromstring(s, uint16).astype(float)
 A *= 1.0/max(A)
 A.shape = w, h
